requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.20
python-telegram-bot==21.9
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
